Moved Date Time mapping to common extensions
