+++ src/Plugins/Widgets.Slider/Extensions.cs
@@ -5,6 +5,8 @@ using System.Reflection;
 using Widgets.Slider.Domain;
 using Widgets.Slider.Models;
 using Grand.Business.Core.Interfaces.Common.Directory;
+using Grand.Web.Common.Extensions;
+
 namespace Widgets.Slider
 {
     public static class MyExtensions
@@ -66,46 +68,5 @@ namespace Widgets.Slider
         {
             return obj.GetProperty(propertyName) != null;
         }
-    }
-
-    public static class DateTimeMappingExtensions
-    {
-        public static DateTime? ConvertToUserTime(this DateTime? datetime, IDateTimeService dateTimeService)
-        {
-            if (datetime.HasValue)
-            {
-                if (datetime.Value.Kind == DateTimeKind.Utc)
-                    datetime = dateTimeService.ConvertToUserTime(datetime.Value, TimeZoneInfo.Utc, dateTimeService.CurrentTimeZone);
-            }
-            return datetime;
-        }
-
-        public static DateTime? ConvertToUtcTime(this DateTime? datetime, IDateTimeService dateTimeService)
-        {
-            if (datetime.HasValue)
-            {
-                if (datetime.Value.Kind != DateTimeKind.Utc)
-                    datetime = dateTimeService.ConvertToUtcTime(datetime.Value, dateTimeService.CurrentTimeZone);
-            }
-            return datetime;
-        }
-
-        public static DateTime ConvertToUserTime(this DateTime datetime, IDateTimeService dateTimeService)
-        {
-            if (datetime.Kind == DateTimeKind.Utc)
-                return dateTimeService.ConvertToUserTime(datetime, TimeZoneInfo.Utc, dateTimeService.CurrentTimeZone);
-
-            return datetime;
-        }
-
-        public static DateTime ConvertToUtcTime(this DateTime datetime, IDateTimeService dateTimeService)
-        {
-            if (datetime.Kind == DateTimeKind.Local)
-                return dateTimeService.ConvertToUtcTime(datetime, dateTimeService.CurrentTimeZone);
-
-            return datetime;
-        }
-    }
-
-
+    }    
 }
